refactor:racingCars에 우승자 구하도록 위임
