+++ src/main/java/nextstep/step3/domain/RacingCars.java
@@ -1,5 +1,6 @@
 package nextstep.step3.domain;
 
+import java.util.Comparator;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -14,5 +15,16 @@ public class RacingCars {
 
 	public List<RacingCar> getRacingCars() {
 		return racingCars;
+	}
+
+	public List<String> getRacingWinnerNames() {
+		int maxPoint = racingCars.stream()
+				.max(Comparator.comparing(RacingCar::getPoint))
+				.get().getPoint();
+
+		return racingCars.stream()
+				.filter(car -> maxPoint == car.getPoint())
+				.map(RacingCar::getName)
+				.collect(Collectors.toList());
 	}
 }
